requirements.txt
absl-py==0.6.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.7.1
backcall==0.1.0
bleach==3.1.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.3.0
defer==1.0.6
defusedxml==0.5.0
dill==0.2.8.2
distro-info==0.18
entrypoints==0.3
future==0.17.1
gast==0.2.2
grpcio==1.17.1
gym==0.10.9
h5py==2.9.0
httplib2==0.9.2
idna==2.6
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.2
mistune==0.8.4
mock==2.0.0
more-itertools==5.0.0
nbconvert==5.4.0
nbformat==4.4.0
netifaces==0.10.4
networkx==2.2
notebook==5.7.4
numpy==1.16.0
oauth==1.0.1
olefile==0.45.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.1
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydot==1.4.1
pyglet==1.3.2
Pygments==2.3.1
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.3.1
pyRFC3339==1.0
python-apt==1.6.3
python-dateutil==2.7.5
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.3
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.20.2
scipy==1.2.0
screen-resolution-extra==0.0.0
seaborn==0.9.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.12.0
sklearn==0.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
torch==1.0.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.29.1
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.2.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
xkit==0.0.0
zope.interface==4.3.2
